// src/app/syllogimous/models/syllogimous.models.ts
export enum EnumScreens {
    Intro = "Intro",
    Start = "Start",
    Tutorial = "Tutorial",
    Game = "Game",
    Feedback = "Feedback",
    History = "History",
    Tutorials = "Tutorials",
    Stats = "Stats",
}

export enum EnumTiers {
    Adept = "Adept",
    Savant = "Savant",
    Mastermind = "Mastermind",
    Visionary = "Visionary",
    Genius = "Genius",
    Luminary = "Luminary",
    Prodigy = "Prodigy",
    Oracle = "Oracle",
    Sage = "Sage",
    Philosopher = "Philosopher",
    Mystic = "Mystic",
    Transcendent = "Transcendent",
}